Changed from float64 to float32. Seems to decrease by 33% the db size :)

+++ src/bgpe/io/readsdssinput.py
@@ -8,5 +8,5 @@ import numpy as np
 
 def Read7xt(arq_obs):
     print '@@@> Reading Starlight Input file: ', arq_obs
-    dt = np.dtype ([('wl', 'float64'), ('flux', 'float64'), ('err', 'float64'), ('flag', 'float64')])
+    dt = np.dtype ([('wl', 'float32'), ('flux', 'float32'), ('err', 'float32'), ('flag', 'float32')])
     return np.loadtxt(arq_obs, dtype = dt)
